+++ lib/ogn_client/messages/receiver.rb
@@ -19,34 +19,34 @@ module OGNClient
           (?<rf_correction_manual>[+-][\d]+)
           (?<rf_correction_automatic>[+-][\d.]+)ppm/
         )?
-        (?<signal>[+-][\d.]+)dB
-        (?:/(?<senders_signal>[+-][\d.]+)dB@10km\[(?<senders_messages>\d+)\])?
-        (?:/(?<good_senders_signal>[+-][\d.]+)dB@10km\[(?<good_senders>\d+)/(?<good_and_bad_senders>\d+)\])?
+        (?<signal_quality>[+-][\d.]+)dB
+        (?:/(?<senders_signal_quality>[+-][\d.]+)dB@10km\[(?<senders_messages>\d+)\])?
+        (?:/(?<good_senders_signal_quality>[+-][\d.]+)dB@10km\[(?<good_senders>\d+)/(?<good_and_bad_senders>\d+)\])?
       )?
     $)x
 
     SUPPORTED_RECEIVER_VERSION = Gem::Version.new('0.2.5')
 
-    attr_reader :version                   # software version as #<Gem::Version "major.minor.patch">
-    attr_reader :platform                  # e.g. "ARM"
-    attr_reader :cpu_load                  # as reported by "uptime"
-    attr_reader :cpu_temperature           # degrees celsius
-    attr_reader :ram_free                  # megabytes
-    attr_reader :ram_total                 # megabytes
-    attr_reader :ntp_offset                # milliseconds
-    attr_reader :ntp_correction            # parts-per-million
-    attr_reader :voltage                   # board voltage in V
-    attr_reader :amperage                  # board amperage in A
-    attr_reader :rf_correction_manual      # as per configuration
-    attr_reader :rf_correction_automatic   # based on GSM
-    attr_reader :senders                   # number of senders within the last hour
-    attr_reader :visible_senders           # number of visible senders within the last hour
-    attr_reader :signal                    # signal-to-noise ratio in decibel
-    attr_reader :senders_signal            # average signal-to-noise ratio in decibel across all senders
-    attr_reader :senders_messages          # number of messages analyzed to calculate the above
-    attr_reader :good_senders_signal       # average signal-to-noise ratio in decibel of good senders (transmitting properly) within the last 24 hours
-    attr_reader :good_and_bad_senders      # number of good and bad senders within the last 24 hours
-    attr_reader :good_senders              # number of good senders (transmitting properly) within the last 24 hours
+    attr_reader :version                       # software version as #<Gem::Version "major.minor.patch">
+    attr_reader :platform                      # e.g. "ARM"
+    attr_reader :cpu_load                      # as reported by "uptime"
+    attr_reader :cpu_temperature               # degrees celsius
+    attr_reader :ram_free                      # megabytes
+    attr_reader :ram_total                     # megabytes
+    attr_reader :ntp_offset                    # milliseconds
+    attr_reader :ntp_correction                # parts-per-million
+    attr_reader :voltage                       # board voltage in V
+    attr_reader :amperage                      # board amperage in A
+    attr_reader :rf_correction_manual          # as per configuration
+    attr_reader :rf_correction_automatic       # based on GSM
+    attr_reader :senders                       # number of senders within the last hour
+    attr_reader :visible_senders               # number of visible senders within the last hour
+    attr_reader :signal_quality                # signal-to-noise ratio in decibel
+    attr_reader :senders_signal_quality        # average signal-to-noise ratio in decibel across all senders
+    attr_reader :senders_messages              # number of messages analyzed to calculate the above
+    attr_reader :good_senders_signal_quality   # average signal-to-noise ratio in decibel of good senders (transmitting properly) within the last 24 hours
+    attr_reader :good_and_bad_senders          # number of good and bad senders within the last 24 hours
+    attr_reader :good_senders                  # number of good senders (transmitting properly) within the last 24 hours
 
     def invisible_senders
       senders - visible_senders
@@ -65,7 +65,7 @@ module OGNClient
     def parse(raw)
       raw.match RECEIVER_PATTERN do |match|
         super unless @raw
-        %i(version platform cpu_load cpu_temperature ram_free ram_total ntp_offset ntp_correction voltage amperage rf_correction_manual rf_correction_automatic senders visible_senders signal senders_signal senders_messages good_senders_signal good_and_bad_senders good_senders).each do |attr|
+        %i(version platform cpu_load cpu_temperature ram_free ram_total ntp_offset ntp_correction voltage amperage rf_correction_manual rf_correction_automatic senders visible_senders signal_quality senders_signal_quality senders_messages good_senders_signal_quality good_and_bad_senders good_senders).each do |attr|
           send("#{attr}=", match[attr]) if match[attr]
         end
         self
@@ -130,20 +130,20 @@ module OGNClient
       @visible_senders = raw.to_i
     end
 
-    def signal=(raw)
-      @signal = raw.to_f.round(3)
+    def signal_quality=(raw)
+      @signal_quality = raw.to_f.round(3)
     end
 
-    def senders_signal=(raw)
-      @senders_signal = raw.to_f.round(3)
+    def senders_signal_quality=(raw)
+      @senders_signal_quality = raw.to_f.round(3)
     end
 
     def senders_messages=(raw)
       @senders_messages = raw.to_i
     end
 
-    def good_senders_signal=(raw)
-      @good_senders_signal = raw.to_f.round(3)
+    def good_senders_signal_quality=(raw)
+      @good_senders_signal_quality = raw.to_f.round(3)
     end
 
     def good_and_bad_senders=(raw)