+++ spec/lib/ogn_client/messages/sender_spec.rb
@@ -15,14 +15,14 @@ describe OGNClient::Sender do
     subject.climb_rate.must_equal 0.1
     subject.turn_rate.must_equal 0.0
     subject.flight_level.must_equal 1.03
-    subject.signal.must_equal 55.2
+    subject.signal_quality.must_equal 55.2
     subject.errors.must_equal 0
     subject.frequency_offset.must_equal -6.2
     subject.gps_accuracy.must_equal [4, 6]
     subject.flarm_software_version.must_equal Gem::Version.new('6.01')
     subject.flarm_hardware_version.must_equal Gem::Version.new(0x03)
     subject.flarm_id.must_equal 'DDACC4'
-    subject.power.must_equal 5.0
+    subject.signal_power.must_equal 5.0
     subject.proximity.must_equal ["D7EA", "DA95"]
   end
 
@@ -86,11 +86,11 @@ describe OGNClient::Sender do
     subject.flarm_id.must_be_nil
   end
 
-  it "must parse valid raw message without power" do
+  it "must parse valid raw message without signal_power" do
     raw = "FLRDF0A52>APRS,qAS,LSTB:/220132h4658.70N/00707.72Ez090/054/A=001424 !W37! id06DF0A52 +020fpm +0.0rot FL001.03 55.2dB 0e -6.2kHz gps4x6 s6.01 h03 rDDACC4 hearD7EA hearDA95"
     subject = OGNClient::Message.parse raw
     subject.must_be_instance_of OGNClient::Sender
-    subject.power.must_be_nil
+    subject.signal_power.must_be_nil
   end
 
   it "must parse valid raw message without proximity" do